Refactor input post-processing for iteration
The previous implementation did not allow a flexible and unified
post-processing framework for parameter specification carried out after
the "main" initial pulse sequence spec, eg for iterating variables.

Refactoring the post-processing functionality makes this possible, by
allowing the post-processing action to act on incomplete sets of
parameter specs. This means that the function can be invoked to
calculate just the iteration start and stop variables before they are
updated.

An additional positive side effect is that redundant specifications can
now be removed from the input string. The responsibility then rests on
the experimenter to ensure that the reference config file says what they
want!

Note that specification of one of the additional or pulse-specified
parameters is currently not implemented and will result in an error
message.