Add incrementation for filenames ending in numeric id
If the script encounters a filename that already exists, but ends in a numeric sequential id, it will attempt to increment it until it finds a valid filename (ie one where the file does not already exist). This can either be done automatically or through user input.

Known bugs: does not work quite as desired for file names specified with no extension... to be fixed soon!