Fix ambiguous behaviour with file extensions
Fixes a bug whereby providing a not-kosher file extension (ie other than ".hdf5") would break the writesafe-check, as Labber will always append a ".hdf5" extension to output files even if this is not provided explicitly. Checking for an existing file now begins with appending the correct extension (by default, ".hdf5") to the path _before_ the file is checked for existence.

Note that if you put multiple extensions on the input filename, it is not my responsibility!