Implement variable relations
A new class, Hdf5Editor, handles directly editing the hdf5 config file in order to implement variable relations. The channels available as equation variables must first be added to the Hdf5Editor instance, after which they can be used in formulas specified through the corresponding class methods.

Currently, the Hdf5Editor edits the *input* reference config file; this will be changed as soon as possible such that the input reference config can remain unchanged as a template for future experiments.

There is also an issue where measurements subsequent to a measurement which initializes a channel relation for a channel cannot then reset the channel to either point values or a sweep (ie it must remain a relation-based channel). This will be fixed by a successful implementation which modifies the output config (instead of the input config, cf above).