Add FileManager class
The FileManager class handles keeping track of the input and output config file names and paths, including tasks such as:
- Making and keeping track of a temporary copy of the original ("template") config file, which should not be modified, such that the copy can be modified by the Hdf5Editor class.
- Testing for existence of the specified output file. (migrated from the InputStrParser.set_MeasurementObject method, but retaining most of the external functions)

At present, FileManager methods such as clean() must be called explicitly; in the future, these should be part of the class constructor/destructor structure. In addition, errors and/or warnings in a few methods could be implemented more cleanly with proper exception handling.