Finish testing Manual instrument
Turns out we just need to add 10 allowed Manual channels to the RHS. We keep the generic channel names, and give them pseudo-descriptive symbols to be used in the channel relation equations.

Note that a proper and thorough treatment of a custom Manual instrument is planned, but it will have to wait for the upcoming unified implementation framework.