Add renaming of copied script
A new option in the script-rcfile enables the name of the script copy (made post-measurement) to match the name of the (final) output database file, with a postfix appended (also specified in the script-rcfile).

The script-rcfile is now, after import by the psictUIFInterface, passed to the FileManager so that it can have direct parameter access (instead of always having them passed as args to methods).

Finally, moved the implementation of the post-measurement copying method into the FileManager object.