Add robust pulse getting and setting
Individual pulses can now be robustly addressed via either a string (in which case it is taken to be the pulse name) or a numerical index (not checked).

The PulseSeq add_pulse method now fully supports both addition of an existing pulse, as well as implicit creation through passing and name and an (optional) parameter dict. Pulse names must be unique, and a KeyError is raised if the new pulse name matches one of the existing pulses.