Add Pulse rcfile
The Pulse rcfile defines the key quantities (and, soon, the pre-processing operations as well as parameters for relations between pulses which will be used to convert between input and output pulse sequences) that are relevant as attributes/parameters of individual pulses in the pulse sequence. For the moment, a few basic ones have been set up to test the functionality.

Note that assigning parameter values to a pulse now sorts them between physical and ordering based on the lists defined in the Pulse rcfile.