Implement pulse sorting
Pulse sorting is carried out by the pulse_sort_attr attribute, as specified in the Pulse rcfile (this should probably not be changed!). A RuntimeError is raised if not all pulses have the absolute_time attribute set; it will be the responsibility of pre-sorting methods to ensure that all the proper calculations are carried out such that this attribute is set for all pulses by the time the sorting method is called.