Improve pulse absolute time calculations
Pulses with (potentially complex) timing dependencies set as 'relative'
or 'previous' will have their times calculated more robustly, as the
pulse list is looped through, skipping pulses that cannot be calculated
in that iteration.

Note that this looping will occur for a maximum number of times equal to
the number of pulses, as we expect at least one pulse to have its time
calculated each iteration. The loop has an overall break condition set
for when all the pulses have a valid absolute time.