Generate complement of inverted pulses
Pulses marked as "is_inverted" are intended to be used to define the complement of a background pulse keeping the qubit in the excited state, except for when we specify it to be coupled with the magnon/cavity. For this, the actual pulses are the complements of the inverted pulses; this commit generates the complementary pulses after absolute_time calculations but before pre-conversion sorting.

Note that the inverted/complementary pulses inherit most of their parameters identically (set as global "inverted" parameters), and differ purely in terms of plateau length and (obviously) time specification.