Add IterationSpec class
Objects of this class can be assigned as parameter values (instead of eg numbers or strings). The LabberExporter will then update the values correctly.

Note that there are many complicated ways in which these can be incorporated into the whole structure and related to the point values and potentially even the channel relations. We will keep it simple for now until specific use cases become clearer.