Modify IterationSpec arithmetic to use only start_time
Only the start_time parameter is used to carry out arithmetic operations with IterationSpec instances; this is to prevent accidental creation of additional iterated parameters when the pulse sequence is being IO-converted.

The intention is to re-introduce this later in such a way that setting iteration parameters with dependencies automatically induces channel relations with those dependencies; however, this may or may not be a desired option.