Implement exporting iteration order
When setting iteration specifications through the Labber API, there is no way to control the order over which multiple variables are iterated, as they will always default to their reference file order. In order to be able to modify this from the script, the order is edited directly in the hdf5 file (after all the specifications have been applied).