Correct pathing for script-rcfile
The script-rcfile can now be specified as a relative path (this is almost essential to having consistent behaviour re-running auto-copied scripts). The copied script files can now be re-run as expected (note that re-runs will, as they should, appear as completely new results with appropriately-incremented filenames), and it will fetch the (correctly-copied) script-rcfile.