Fix implicit calculation of pulse sequence length
The script's implicit calculation of the "Number of points" parameter was not correctly taking into account the width and truncation range of the final pulse. This is now fixed, so the SQPG should output exactly the right number of points required for the pulse sequence.

Note that if the total length of the pulse sequence is changed through iterations or relations, a relation must still be explicitly set.