Move copying to before measurement
If the script file is edited before the measurement is completed (which is often done as we prepare the next measurement in advance), this "mismatched" script file will be the one copied over. As the copying does not require any other files (just script and script-rc), we can just do it before the measurement is carried out. In this way, having Labber open and carrying out a measurement guarantees that the script has already been copied, and so the file can safely be modified in preparation for the next experiment.