Add optional dead time to SQPG main parameters
We can eliminate the need to create an explicit pulse for including dead time after a pulse sequence if we just extend the SQPG "Number of points" parameter. This is now implemented by specifying a "dead_time" in the "main" SQPG parameters (the different name format is chosen to emphasize the fact that it is not a true SQPG parameter, and is not set through Labber). The dead time will default to 0.0 if it is not specified, so can be ignored if it is not needed or specified explicitly in some other way (eg inverted pulse sequences).