Add slave status
When running as a "slave" (ie part of an automated workflow with "master" scripts), the integrated script-copying routine fails (as the master script is copied instead of the actual measurement scripts). It is now possible to specify a slave status, which, when true, disables the integrated script copying. The onus is then on the master script to carry out the script copying.

In the future, this will either be integrated more comprehensively as an additional module for such master scripts, or by creating a robust means of identifying the measurement script within the PSICT framework.