Add pulse sequence length calculations
The 'Number of points' parameter of the SQPG is now calculated based on the two parameters 'sequence_duration' and 'Sample rate'. Previously, this was carried out by setting a relation explicitly in the pulse sequence (note that this approach is still backwards-compatible), but presently if the two mentioned parameters are specified in the point values dict, the 'Number of points' will be automatically calculated as their product.