Add functionality to set parameters through InstrumentClient interface
For some reason, it appears as though it is not possible to set input values as lists/arrays through either the Labber scripting API or by direct editing of the hdf5 file. Rather oddly, it is almost trivial to do so through the Labber InstrumentClient interface (not part of the ScriptTools package, but the top-level Labber package!). This commit adds this compatibility.