Remove superfluous quantities from MultiPulse
After testing, it looks like the best thing to do is to simply outsource both the pulse definitions and pulse sequences to external files. Optimally, these can be for all intents and purposes temporary, created by the measurement script itself.

Currently, the functions to write these files are contained directly in the external script, but once they are verified to work as intended, they will be moved to their own module in the PSICT_extras foider.