Add preliminary pulse sequence calculations
The 'Pulse counter' is now used to specify which pulse sequence is to be used in a given iteration (allowing for in-Labber iteration through the values). Additionally, preliminary calculations of total time and number of points of the chosen pulse sequence have been added. [In the future, we should definitely add an option to specify a fixed total time independently of the pulse sequences. However, it may be instructive to see how Labber behaves with a variable time for the moment...]