Add reverse waveform generation from final pulse
If generation from final pulse is set, the waveform is generated in reverse, with the final pulse occurring at the specified time. This is useful for eg keeping the readout pulse at the same position in the pulse sequence every time, regardless of the timing of pulses that come before it.