Skip API-client value setting on server connection error
Dry-runs without a server connection were previously impossible, as the LabberExporter would always attempt to connect to the server irrespectively of whether any values were actually being set through the API-client or not. As this feature is a bit problematic anyway, it has just been set to silently pass over if no connection to the server can be made. (Probably not the best to do a catch-all exception, but it will have to do for now...)