Move config loading to psictUIFInterface creation
In order to allow the config file to store user-specified logging config options, the config file loading needs to be carried out as early as possible in the psictUIFInterface instance initialization process.

And unfortunate downside is that the config loading process itself now cannot be logged; however, the config file path will naturally be logged post-fact, thereby still allowing some degree of debugging.