Move automation script into PSICT
Previously, higher-level automation (ie a 'master' script which allowed control over multiple experiments being run in sequence, each corresponding to a separate and self-contained execution of a single experiment script) was carried out by a module outside of the PSICT scope. For completeness, as well as a secondary consequence of improving the backend stability and logging, this has now been moved into the `PSICT_extras` folder, named `WorkerScriptManager`.

In addition, in order to allow more robust debugging and tracking new features, test scripts will now be included in the `sample` directory.