Pass parent logger name to psictUIFInterface instance
The WorkerScriptManager now automatically sets the parent logger name to 'WorkerScriptManager' just before the measurement is run. The worker script has this as an option in the `PSICT_options`, but it is set to `None` in the main-only segment (such that it is always set to None before running as a standalone).