Add logging initialization in WorkerScriptManager
Now that there were no obstacles remaining, it was possible to implement proper logging in the WorkerScriptManager class. This follows the format of the psictUIFInterface loggers, and re-uses the logging initialization code.

Note that one of the relatively-major changes is to the call signature of the WorkerScriptManager init method: all logging-related args have been removed (the information is taken from the specified PSICT config file), but now naturally the PSICT config file must be specified in addition to the worker script.