Add log method to psictUIFInterface class
This commit adds a 'log' method in the psictUIFInterface class that allows for quick and painless logging of events from the worker script so that they are retained in log files for future reference.

Note that the arg order is reversed from the usual logger.log() call to facilitate 'instant printing'. If more configurability is desired, additional arguments can be passed directly as though to logger.log().