Fix rounding errors for subsampled MultiPulse sequences
When speciying timing with sub-sample-rate precision, the MultiPulse algorithm (as it is based on the SQPG algorithm) would cause rounding errors to unintentionally modify the sequence timing. This is now fixed by changing the 'head' of the pulse sequence to be based on the exact time, instead of the sample-rate-based time index. tl;dr the rounding of the timing to the sample rate is now done later in the pulse generation process.