Allow disabling reference file copy deletion
During testing, it is sometimes necessary to examine the (temporary) reference file copy to which PSICT actually applies parameter values. This is especially important when testing outside of a Labber-licensed environment, as it is not possible to actually create an output file.

Previously, the reference file copy ws always deleted. This commit introduces a parameter in the FileManager rcfile (DEL_REF_COPY) which can be set to False, thus preventing the automatic deletion of the reference file copy.