Add logfile comment functionality
These additions make it possible to set the logfile comment text (typically viewed in the Log Browser) in the reference file copy, and therefore the output file as well.

Previously, this comment was fixed based on what was already present in the original template file at the time PSICT was run. Now, it can be dynamically adjusted and is fully compatibler with the WorkerScriptManager automation paradigm: it is simply an additional key-value pair in the `PSICT_options` dict in the sample script framework, and has its own UIF interface setting method (also shown in the sample script).